use nn.Sequential to define network architecture for training
Neet to be caution with the loss function and the corresponding models
e.g. if using nll_loss, then need to have LogSoftmax at the end of model

Also, currently when the batch size is not a multiple of 10 (say 64),
there will be errors during training. Need to fix this.